+++ src/ModuleInstance/ModuleInstance.php
@@ -10,6 +10,7 @@ use BristolSU\Support\ModuleInstance\Connection\ModuleInstanceService;
 use BristolSU\Support\ModuleInstance\Contracts\ModuleInstance as ModuleInstanceContract;
 use BristolSU\Support\Permissions\Models\ModuleInstancePermission;
 use BristolSU\Support\ModuleInstance\Settings\ModuleInstanceSetting;
+use Illuminate\Database\Eloquent\Builder;
 use Illuminate\Database\Eloquent\Model;
 use Illuminate\Database\Eloquent\ModelNotFoundException;
 use Illuminate\Support\Str;
@@ -175,5 +176,10 @@ class ModuleInstance extends Model implements ModuleInstanceContract
         } catch (ModelNotFoundException $e) {
             return $default;
         }
+    }
+
+    public function scopeEnabled(Builder $query)
+    {
+        return $query->where('enabled', true);
     }
 }